+++ lib/primitive/cardano-wallet-primitive.cabal
@@ -167,30 +167,30 @@ library
     Cardano.Wallet.Primitive.Types.UTxO.Gen
     Cardano.Wallet.Primitive.Types.ValidityIntervalExplicit
     Cardano.Wallet.Primitive.Types.WitnessCount
-    Cardano.Wallet.Read.Primitive.Block
-    Cardano.Wallet.Read.Primitive.Block.Header
-    Cardano.Wallet.Read.Primitive.Tx
-    Cardano.Wallet.Read.Primitive.Tx.Allegra
-    Cardano.Wallet.Read.Primitive.Tx.Alonzo
-    Cardano.Wallet.Read.Primitive.Tx.Babbage
-    Cardano.Wallet.Read.Primitive.Tx.Byron
-    Cardano.Wallet.Read.Primitive.Tx.Conway
-    Cardano.Wallet.Read.Primitive.Tx.Features.Certificates
-    Cardano.Wallet.Read.Primitive.Tx.Features.CollateralInputs
-    Cardano.Wallet.Read.Primitive.Tx.Features.CollateralOutputs
-    Cardano.Wallet.Read.Primitive.Tx.Features.ExtraSigs
-    Cardano.Wallet.Read.Primitive.Tx.Features.Fee
-    Cardano.Wallet.Read.Primitive.Tx.Features.Inputs
-    Cardano.Wallet.Read.Primitive.Tx.Features.Integrity
-    Cardano.Wallet.Read.Primitive.Tx.Features.Metadata
-    Cardano.Wallet.Read.Primitive.Tx.Features.Mint
-    Cardano.Wallet.Read.Primitive.Tx.Features.Outputs
-    Cardano.Wallet.Read.Primitive.Tx.Features.ScriptValidity
-    Cardano.Wallet.Read.Primitive.Tx.Features.Validity
-    Cardano.Wallet.Read.Primitive.Tx.Features.Withdrawals
-    Cardano.Wallet.Read.Primitive.Tx.Mary
-    Cardano.Wallet.Read.Primitive.Tx.Sealed
-    Cardano.Wallet.Read.Primitive.Tx.Shelley
+    Cardano.Wallet.Primitive.Ledger.Read.Block
+    Cardano.Wallet.Primitive.Ledger.Read.Block.Header
+    Cardano.Wallet.Primitive.Ledger.Read.Tx
+    Cardano.Wallet.Primitive.Ledger.Read.Tx.Allegra
+    Cardano.Wallet.Primitive.Ledger.Read.Tx.Alonzo
+    Cardano.Wallet.Primitive.Ledger.Read.Tx.Babbage
+    Cardano.Wallet.Primitive.Ledger.Read.Tx.Byron
+    Cardano.Wallet.Primitive.Ledger.Read.Tx.Conway
+    Cardano.Wallet.Primitive.Ledger.Read.Tx.Features.Certificates
+    Cardano.Wallet.Primitive.Ledger.Read.Tx.Features.CollateralInputs
+    Cardano.Wallet.Primitive.Ledger.Read.Tx.Features.CollateralOutputs
+    Cardano.Wallet.Primitive.Ledger.Read.Tx.Features.ExtraSigs
+    Cardano.Wallet.Primitive.Ledger.Read.Tx.Features.Fee
+    Cardano.Wallet.Primitive.Ledger.Read.Tx.Features.Inputs
+    Cardano.Wallet.Primitive.Ledger.Read.Tx.Features.Integrity
+    Cardano.Wallet.Primitive.Ledger.Read.Tx.Features.Metadata
+    Cardano.Wallet.Primitive.Ledger.Read.Tx.Features.Mint
+    Cardano.Wallet.Primitive.Ledger.Read.Tx.Features.Outputs
+    Cardano.Wallet.Primitive.Ledger.Read.Tx.Features.ScriptValidity
+    Cardano.Wallet.Primitive.Ledger.Read.Tx.Features.Validity
+    Cardano.Wallet.Primitive.Ledger.Read.Tx.Features.Withdrawals
+    Cardano.Wallet.Primitive.Ledger.Read.Tx.Mary
+    Cardano.Wallet.Primitive.Ledger.Read.Tx.Sealed
+    Cardano.Wallet.Primitive.Ledger.Read.Tx.Shelley
     Cardano.Wallet.Unsafe
     Cardano.Wallet.Util
     Control.Monad.Random.NonRandom
