+++ lib/primitive/lib/Cardano/Wallet/Primitive/Ledger/Read/Tx/Allegra.hs
@@ -1,23 +1,34 @@
+{-# LANGUAGE AllowAmbiguousTypes #-}
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE DuplicateRecordFields #-}
 {-# LANGUAGE FlexibleContexts #-}
+{-# LANGUAGE FlexibleInstances #-}
+{-# LANGUAGE GADTs #-}
+{-# LANGUAGE QuantifiedConstraints #-}
+{-# LANGUAGE RankNTypes #-}
+{-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TupleSections #-}
+{-# LANGUAGE TypeFamilies #-}
+{-# LANGUAGE UndecidableInstances #-}
 
 -- |
 -- Copyright: © 2020-2022 IOHK
 -- License: Apache-2.0
 --
 
-module Cardano.Wallet.Read.Primitive.Tx.Mary
-    ( fromMaryTx
-    , fromMaryTx'
+module Cardano.Wallet.Primitive.Ledger.Read.Tx.Allegra
+    ( fromAllegraTx
+    , fromAllegraTx'
     )
     where
 
 import Prelude
 
+import Cardano.Address.Script
+    ( KeyRole (..)
+    )
 import Cardano.Api
-    ( MaryEra
+    ( AllegraEra
     )
 import Cardano.Ledger.Api
     ( addrTxWitsL
@@ -27,14 +38,36 @@ import Cardano.Ledger.Api
     , certsTxBodyL
     , feeTxBodyL
     , inputsTxBodyL
-    , mintTxBodyL
     , outputsTxBodyL
     , scriptTxWitsL
     , vldtTxBodyL
     , witsTxL
     )
+import Cardano.Ledger.Core
+    ()
+import Cardano.Ledger.Shelley.Tx
+    ( ShelleyTx
+    )
 import Cardano.Wallet.Primitive.Ledger.Convert
     ( toWalletScript
+    )
+import Cardano.Wallet.Primitive.Ledger.Read.Tx.Features.Certificates
+    ( anyEraCerts
+    )
+import Cardano.Wallet.Primitive.Ledger.Read.Tx.Features.Inputs
+    ( fromShelleyTxIn
+    )
+import Cardano.Wallet.Primitive.Ledger.Read.Tx.Features.Metadata
+    ( fromAllegraMetadata
+    )
+import Cardano.Wallet.Primitive.Ledger.Read.Tx.Features.Outputs
+    ( fromAllegraTxOut
+    )
+import Cardano.Wallet.Primitive.Ledger.Read.Tx.Features.Validity
+    ( afterShelleyValidityInterval
+    )
+import Cardano.Wallet.Primitive.Ledger.Read.Tx.Features.Withdrawals
+    ( fromLedgerWithdrawals
     )
 import Cardano.Wallet.Primitive.Types.AnyExplicitScripts
     ( AnyExplicitScript (NativeExplicitScript)
@@ -42,42 +75,20 @@ import Cardano.Wallet.Primitive.Types.AnyExplicitScripts
 import Cardano.Wallet.Primitive.Types.TokenMapWithScripts
     ( ScriptReference (ViaSpending)
     , TokenMapWithScripts
+    , emptyTokenMapWithScripts
     )
 import Cardano.Wallet.Primitive.Types.ValidityIntervalExplicit
     ( ValidityIntervalExplicit
     )
 import Cardano.Wallet.Primitive.Types.WitnessCount
     ( WitnessCount (WitnessCount)
-    , WitnessCountCtx
-    , toKeyRole
     )
 import Cardano.Wallet.Read.Eras
-    ( inject
-    , mary
-    )
-import Cardano.Wallet.Read.Primitive.Tx.Features.Certificates
-    ( anyEraCerts
-    )
-import Cardano.Wallet.Read.Primitive.Tx.Features.Inputs
-    ( fromShelleyTxIn
-    )
-import Cardano.Wallet.Read.Primitive.Tx.Features.Metadata
-    ( fromMaryMetadata
-    )
-import Cardano.Wallet.Read.Primitive.Tx.Features.Mint
-    ( maryMint
-    )
-import Cardano.Wallet.Read.Primitive.Tx.Features.Outputs
-    ( fromMaryTxOut
-    )
-import Cardano.Wallet.Read.Primitive.Tx.Features.Validity
-    ( afterShelleyValidityInterval
-    )
-import Cardano.Wallet.Read.Primitive.Tx.Features.Withdrawals
-    ( fromLedgerWithdrawals
+    ( allegra
+    , inject
     )
 import Cardano.Wallet.Read.Tx
-    ( Tx (Tx)
+    ( Tx (..)
     )
 import Cardano.Wallet.Read.Tx.CBOR
     ( renderTxToCBOR
@@ -93,19 +104,22 @@ import Control.Lens
     , (^.)
     , (^..)
     )
+import Data.Foldable
+    ( toList
+    )
 
 import qualified Cardano.Api.Shelley as Cardano
 import qualified Cardano.Ledger.BaseTypes as SL
-import qualified Cardano.Ledger.Shelley.API as SL
 import qualified Cardano.Wallet.Primitive.Ledger.Convert as Ledger
 import qualified Cardano.Wallet.Primitive.Types.Certificates as W
 import qualified Cardano.Wallet.Primitive.Types.Hash as W
 import qualified Cardano.Wallet.Primitive.Types.Tx as W
 import qualified Data.Set as Set
 
-fromMaryTx
-    :: SL.ShelleyTx (Cardano.ShelleyLedgerEra MaryEra)
-    -> WitnessCountCtx
+-- NOTE: For resolved inputs we have to pass in a dummy value of 0.
+
+fromAllegraTx
+    :: ShelleyTx (Cardano.ShelleyLedgerEra AllegraEra)
     -> ( W.Tx
        , [W.Certificate]
        , TokenMapWithScripts
@@ -113,46 +127,41 @@ fromMaryTx
        , Maybe ValidityIntervalExplicit
        , WitnessCount
        )
-fromMaryTx tx witCtx =
-    ( fromMaryTx' tx
-    , anyEraCerts $ tx ^. bodyTxL.certsTxBodyL
-    , assetsToMint
-    , assetsToBurn
+fromAllegraTx tx =
+    ( fromAllegraTx' tx
+    , anyEraCerts $ tx ^. bodyTxL . certsTxBodyL
+    , emptyTokenMapWithScripts
+    , emptyTokenMapWithScripts
     , Just $ afterShelleyValidityInterval $ tx ^. bodyTxL.vldtTxBodyL
     , WitnessCount
         (fromIntegral $ Set.size $ tx ^. witsTxL.addrTxWitsL)
         ((`NativeExplicitScript` ViaSpending)
-         . toWalletScript (toKeyRole witCtx)
+         . toWalletScript (\_vkey -> Payment)
             <$> tx ^.. witsTxL.scriptTxWitsL.folded)
         (fromIntegral $ Set.size $ tx ^. witsTxL.bootAddrTxWitsL)
     )
-  where
-    (assetsToMint, assetsToBurn) =
-        maryMint (tx ^. bodyTxL.mintTxBodyL) (tx ^. witsTxL)
 
-fromMaryTx' ::
-    SL.ShelleyTx (Cardano.ShelleyLedgerEra MaryEra) ->
-    W.Tx
-fromMaryTx' tx =
+fromAllegraTx' :: ShelleyTx (Cardano.ShelleyLedgerEra AllegraEra) -> W.Tx
+fromAllegraTx' tx =
     W.Tx
         { txId =
             W.Hash $ shelleyTxHash tx
         , txCBOR =
-            Just $ renderTxToCBOR $ inject mary $ Tx tx
+            Just $ renderTxToCBOR $ inject allegra $ Tx tx
         , fee =
             Just $ Ledger.toWalletCoin $ tx ^. bodyTxL . feeTxBodyL
         , resolvedInputs =
             (,Nothing) . fromShelleyTxIn <$> tx ^.. bodyTxL . inputsTxBodyL . folded
         , resolvedCollateralInputs =
-            []
+            [] -- TODO: (ADP-957)
         , outputs =
-            fromMaryTxOut <$> tx ^.. bodyTxL . outputsTxBodyL . folded
+            fromAllegraTxOut <$> toList (tx ^. bodyTxL . outputsTxBodyL)
         , collateralOutput =
-            Nothing -- Collateral outputs are not supported in Mary.
+            Nothing -- Collateral outputs are not supported in Allegra.
         , withdrawals =
             fromLedgerWithdrawals . shelleyWithdrawals $ tx
         , metadata =
-            fromMaryMetadata <$> SL.strictMaybeToMaybe (tx ^. auxDataTxL)
+            fromAllegraMetadata <$> SL.strictMaybeToMaybe (tx ^. auxDataTxL)
         , scriptValidity =
             Nothing
         }
