+++ lib/primitive/lib/Cardano/Wallet/Primitive/Ledger/Read/Tx/Babbage.hs
@@ -10,29 +10,27 @@
 -- License: Apache-2.0
 --
 
-module Cardano.Wallet.Read.Primitive.Tx.Conway
-    ( fromConwayTx
-    , fromConwayTx'
+module Cardano.Wallet.Primitive.Ledger.Read.Tx.Babbage
+    ( fromBabbageTx
+    , fromBabbageTx'
     )
     where
 
 import Prelude
 
 import Cardano.Api
-    ( ConwayEra
-    )
-import Cardano.Ledger.Alonzo.Scripts
-    ( AlonzoScript
+    ( BabbageEra
     )
 import Cardano.Ledger.Api
-    ( StandardCrypto
+    ( ScriptHash
+    , StandardCrypto
     , addrTxWitsL
     , auxDataTxL
     , bodyTxL
     , bootAddrTxWitsL
+    , certsTxBodyL
     , collateralInputsTxBodyL
     , collateralReturnTxBodyL
-    , conwayCertsTxBodyL
     , feeTxBodyL
     , inputsTxBodyL
     , isValidTxL
@@ -43,25 +41,34 @@ import Cardano.Ledger.Api
     , vldtTxBodyL
     , witsTxL
     )
-import Cardano.Ledger.Babbage
-    ( BabbageTxOut
-    )
 import Cardano.Wallet.Primitive.Ledger.Convert
     ( toWalletScript
     , toWalletTokenPolicyId
     )
-import Cardano.Wallet.Primitive.Types.AnyExplicitScripts
-    ( AnyExplicitScript (..)
+import Cardano.Wallet.Primitive.Ledger.Read.Tx.Features.Certificates
+    ( anyEraCerts
     )
-import Cardano.Wallet.Primitive.Types.TokenMapWithScripts
-    ( PlutusScriptInfo (PlutusScriptInfo)
-    , PlutusVersion (PlutusVersionV1, PlutusVersionV2, PlutusVersionV3)
-    , ReferenceInput (ReferenceInput)
-    , ScriptReference (..)
-    , TokenMapWithScripts
+import Cardano.Wallet.Primitive.Ledger.Read.Tx.Features.Inputs
+    ( fromShelleyTxIn
+    )
+import Cardano.Wallet.Primitive.Ledger.Read.Tx.Features.Metadata
+    ( fromBabbageMetadata
+    )
+import Cardano.Wallet.Primitive.Ledger.Read.Tx.Features.Mint
+    ( babbageMint
+    , fromLedgerScriptHash
+    )
+import Cardano.Wallet.Primitive.Ledger.Read.Tx.Features.Outputs
+    ( fromBabbageTxOut
+    )
+import Cardano.Wallet.Primitive.Ledger.Read.Tx.Features.Validity
+    ( afterShelleyValidityInterval
+    )
+import Cardano.Wallet.Primitive.Ledger.Read.Tx.Features.Withdrawals
+    ( fromLedgerWithdrawals
     )
 import Cardano.Wallet.Primitive.Types.TokenPolicy
-    ( TokenPolicyId (..)
+    ( TokenPolicyId
     )
 import Cardano.Wallet.Primitive.Types.Tx
     ( Tx (txId)
@@ -69,39 +76,9 @@ import Cardano.Wallet.Primitive.Types.Tx
 import Cardano.Wallet.Primitive.Types.Tx.TxIn
     ( TxIn (..)
     )
-import Cardano.Wallet.Primitive.Types.ValidityIntervalExplicit
-    ( ValidityIntervalExplicit
-    )
-import Cardano.Wallet.Primitive.Types.WitnessCount
-    ( WitnessCount (WitnessCount)
-    , WitnessCountCtx
-    , toKeyRole
-    )
 import Cardano.Wallet.Read.Eras
-    ( conway
+    ( babbage
     , inject
-    )
-import Cardano.Wallet.Read.Primitive.Tx.Features.Certificates
-    ( fromConwayCerts
-    )
-import Cardano.Wallet.Read.Primitive.Tx.Features.Inputs
-    ( fromShelleyTxIn
-    )
-import Cardano.Wallet.Read.Primitive.Tx.Features.Metadata
-    ( fromConwayMetadata
-    )
-import Cardano.Wallet.Read.Primitive.Tx.Features.Mint
-    ( conwayMint
-    , fromLedgerScriptHash
-    )
-import Cardano.Wallet.Read.Primitive.Tx.Features.Outputs
-    ( fromConwayTxOut
-    )
-import Cardano.Wallet.Read.Primitive.Tx.Features.Validity
-    ( afterShelleyValidityInterval
-    )
-import Cardano.Wallet.Read.Primitive.Tx.Features.Withdrawals
-    ( fromLedgerWithdrawals
     )
 import Cardano.Wallet.Read.Tx
     ( Tx (..)
@@ -115,14 +92,34 @@ import Cardano.Wallet.Read.Tx.Hash
 import Cardano.Wallet.Read.Tx.Withdrawals
     ( shelleyWithdrawals
     )
+
+import Cardano.Ledger.Babbage
+    ( AlonzoScript
+    , BabbageTxOut
+    )
+import Cardano.Wallet.Primitive.Types.AnyExplicitScripts
+    ( AnyExplicitScript (..)
+    )
+import Cardano.Wallet.Primitive.Types.TokenMapWithScripts
+    ( PlutusScriptInfo (PlutusScriptInfo)
+    , PlutusVersion (PlutusVersionV1, PlutusVersionV2, PlutusVersionV3)
+    , ReferenceInput (ReferenceInput)
+    , ScriptReference (ViaReferenceInput, ViaSpending)
+    , TokenMapWithScripts
+    )
+import Cardano.Wallet.Primitive.Types.ValidityIntervalExplicit
+    ( ValidityIntervalExplicit
+    )
+import Cardano.Wallet.Primitive.Types.WitnessCount
+    ( WitnessCount (WitnessCount)
+    , WitnessCountCtx
+    , toKeyRole
+    )
 import Control.Lens
     ( folded
     , (<&>)
     , (^.)
     , (^..)
-    )
-import Data.Foldable
-    ( toList
     )
 import Data.Map
     ( Map
@@ -134,13 +131,12 @@ import Data.Word
     ( Word32
     )
 import Ouroboros.Consensus.Cardano.Block
-    ( StandardConway
+    ( StandardBabbage
     )
 
 import qualified Cardano.Api.Shelley as Cardano
 import qualified Cardano.Ledger.Alonzo.Scripts as Alonzo
 import qualified Cardano.Ledger.Alonzo.Tx as Alonzo
-import qualified Cardano.Ledger.Api as Ledger
 import qualified Cardano.Ledger.BaseTypes as SL
 import qualified Cardano.Ledger.Core as Core
 import qualified Cardano.Ledger.Language as Language
@@ -153,8 +149,8 @@ import qualified Data.List as L
 import qualified Data.Map.Strict as Map
 import qualified Data.Set as Set
 
-fromConwayTx
-    :: Alonzo.AlonzoTx (Cardano.ShelleyLedgerEra ConwayEra)
+fromBabbageTx
+    :: Alonzo.AlonzoTx (Cardano.ShelleyLedgerEra BabbageEra)
     -> WitnessCountCtx
     -> ( W.Tx
        , [W.Certificate]
@@ -163,9 +159,9 @@ fromConwayTx
        , Maybe ValidityIntervalExplicit
        , WitnessCount
        )
-fromConwayTx tx witCtx =
+fromBabbageTx tx witCtx =
     ( tx'
-    , fmap fromConwayCerts . toList $ tx ^. bodyTxL . conwayCertsTxBodyL
+    , anyEraCerts $ tx ^. bodyTxL . certsTxBodyL
     , assetsToMint
     , assetsToBurn
     , Just $ afterShelleyValidityInterval $ tx ^. bodyTxL.vldtTxBodyL
@@ -175,7 +171,7 @@ fromConwayTx tx witCtx =
         (fromIntegral $ Set.size $ tx ^. witsTxL.bootAddrTxWitsL)
     )
   where
-    tx' = fromConwayTx' tx
+    tx' = fromBabbageTx' tx
     txId' = txId tx'
 
     anyScriptsFromTxOuts :: Map TokenPolicyId AnyExplicitScript
@@ -188,16 +184,16 @@ fromConwayTx tx witCtx =
       where
         scriptWithHashIx
             :: Word32
-            -> BabbageTxOut StandardConway
+            -> BabbageTxOut StandardBabbage
             -> Maybe
                 ( ScriptReference
-                , Ledger.ScriptHash StandardCrypto
-                , AlonzoScript StandardConway
+                , ScriptHash StandardCrypto
+                , AlonzoScript StandardBabbage
                 )
         scriptWithHashIx ix txout =
-            snd (fromConwayTxOut txout) <&> \script ->
+            snd (fromBabbageTxOut txout) <&> \script ->
                 ( ViaReferenceInput (ReferenceInput (TxIn txId' ix))
-                , hashConwayScript script
+                , hashBabbageScript script
                 , script
                 )
 
@@ -209,15 +205,15 @@ fromConwayTx tx witCtx =
             ]
 
     (assetsToMint, assetsToBurn) =
-        conwayMint
+        babbageMint
             (tx ^. bodyTxL.referenceInputsTxBodyL)
             (tx ^. bodyTxL.mintTxBodyL)
             (tx ^. witsTxL)
 
     fromLedgerToAnyScript
         :: ( ScriptReference
-           , Ledger.ScriptHash StandardCrypto
-           , AlonzoScript StandardConway
+           , ScriptHash StandardCrypto
+           , AlonzoScript StandardBabbage
            )
         -> (TokenPolicyId, AnyExplicitScript)
     fromLedgerToAnyScript (scriptRef, scriptH, script) =
@@ -232,38 +228,38 @@ fromConwayTx tx witCtx =
                 PlutusExplicitScript
                     (PlutusScriptInfo
                         (toPlutusVer ver)
-                        (fromLedgerScriptHash $ hashConwayScript script))
+                        (fromLedgerScriptHash $ hashBabbageScript script))
                     scriptRef
 
         toPlutusVer Language.PlutusV1 = PlutusVersionV1
         toPlutusVer Language.PlutusV2 = PlutusVersionV2
         toPlutusVer Language.PlutusV3 = PlutusVersionV3
 
-    hashConwayScript = Core.hashScript @(Cardano.ShelleyLedgerEra ConwayEra)
+    hashBabbageScript = Core.hashScript @(Cardano.ShelleyLedgerEra BabbageEra)
 
-fromConwayTx' :: Alonzo.AlonzoTx (Cardano.ShelleyLedgerEra ConwayEra) -> W.Tx
-fromConwayTx' tx =
+fromBabbageTx' :: Alonzo.AlonzoTx (Cardano.ShelleyLedgerEra BabbageEra) -> W.Tx
+fromBabbageTx' tx =
     W.Tx
         { txId = W.Hash $ shelleyTxHash tx
         , txCBOR =
-            Just $ renderTxToCBOR $ inject conway $ Tx tx
+            Just $ renderTxToCBOR $ inject babbage $ Tx tx
         , fee =
-            Just $ Ledger.toWalletCoin $ tx ^. bodyTxL . feeTxBodyL
+            Just $ Ledger.toWalletCoin $ tx ^. bodyTxL.feeTxBodyL
         , resolvedInputs =
             (,Nothing) . fromShelleyTxIn
-                <$> tx ^.. bodyTxL . inputsTxBodyL . folded
+                <$> tx ^.. bodyTxL.inputsTxBodyL.folded
         , resolvedCollateralInputs =
             (,Nothing) . fromShelleyTxIn
-                <$> tx ^.. bodyTxL . collateralInputsTxBodyL . folded
+                <$> tx ^.. bodyTxL.collateralInputsTxBodyL.folded
         , outputs =
-            fst . fromConwayTxOut <$> tx ^.. bodyTxL . outputsTxBodyL . folded
+            fst . fromBabbageTxOut <$> tx ^.. bodyTxL.outputsTxBodyL.folded
         , collateralOutput =
             strictMaybeToMaybe $
-                fst . fromConwayTxOut <$> tx ^. bodyTxL . collateralReturnTxBodyL
+                fst . fromBabbageTxOut <$> tx ^. bodyTxL.collateralReturnTxBodyL
         , withdrawals =
             fromLedgerWithdrawals . shelleyWithdrawals $ tx
         , metadata =
-            fromConwayMetadata <$> SL.strictMaybeToMaybe (tx ^. auxDataTxL)
+            fromBabbageMetadata <$> SL.strictMaybeToMaybe (tx ^. auxDataTxL)
         , scriptValidity =
             Just $ case tx ^. isValidTxL of
                 Alonzo.IsValid True -> W.TxScriptValid
