+++ lib/primitive/lib/Cardano/Wallet/Primitive/Ledger/Read/Tx/Conway.hs
@@ -10,27 +10,29 @@
 -- License: Apache-2.0
 --
 
-module Cardano.Wallet.Read.Primitive.Tx.Babbage
-    ( fromBabbageTx
-    , fromBabbageTx'
+module Cardano.Wallet.Primitive.Ledger.Read.Tx.Conway
+    ( fromConwayTx
+    , fromConwayTx'
     )
     where
 
 import Prelude
 
 import Cardano.Api
-    ( BabbageEra
+    ( ConwayEra
+    )
+import Cardano.Ledger.Alonzo.Scripts
+    ( AlonzoScript
     )
 import Cardano.Ledger.Api
-    ( ScriptHash
-    , StandardCrypto
+    ( StandardCrypto
     , addrTxWitsL
     , auxDataTxL
     , bodyTxL
     , bootAddrTxWitsL
-    , certsTxBodyL
     , collateralInputsTxBodyL
     , collateralReturnTxBodyL
+    , conwayCertsTxBodyL
     , feeTxBodyL
     , inputsTxBodyL
     , isValidTxL
@@ -41,12 +43,47 @@ import Cardano.Ledger.Api
     , vldtTxBodyL
     , witsTxL
     )
+import Cardano.Ledger.Babbage
+    ( BabbageTxOut
+    )
 import Cardano.Wallet.Primitive.Ledger.Convert
     ( toWalletScript
     , toWalletTokenPolicyId
     )
+import Cardano.Wallet.Primitive.Ledger.Read.Tx.Features.Certificates
+    ( fromConwayCerts
+    )
+import Cardano.Wallet.Primitive.Ledger.Read.Tx.Features.Inputs
+    ( fromShelleyTxIn
+    )
+import Cardano.Wallet.Primitive.Ledger.Read.Tx.Features.Metadata
+    ( fromConwayMetadata
+    )
+import Cardano.Wallet.Primitive.Ledger.Read.Tx.Features.Mint
+    ( conwayMint
+    , fromLedgerScriptHash
+    )
+import Cardano.Wallet.Primitive.Ledger.Read.Tx.Features.Outputs
+    ( fromConwayTxOut
+    )
+import Cardano.Wallet.Primitive.Ledger.Read.Tx.Features.Validity
+    ( afterShelleyValidityInterval
+    )
+import Cardano.Wallet.Primitive.Ledger.Read.Tx.Features.Withdrawals
+    ( fromLedgerWithdrawals
+    )
+import Cardano.Wallet.Primitive.Types.AnyExplicitScripts
+    ( AnyExplicitScript (..)
+    )
+import Cardano.Wallet.Primitive.Types.TokenMapWithScripts
+    ( PlutusScriptInfo (PlutusScriptInfo)
+    , PlutusVersion (PlutusVersionV1, PlutusVersionV2, PlutusVersionV3)
+    , ReferenceInput (ReferenceInput)
+    , ScriptReference (..)
+    , TokenMapWithScripts
+    )
 import Cardano.Wallet.Primitive.Types.TokenPolicy
-    ( TokenPolicyId
+    ( TokenPolicyId (..)
     )
 import Cardano.Wallet.Primitive.Types.Tx
     ( Tx (txId)
@@ -54,31 +91,17 @@ import Cardano.Wallet.Primitive.Types.Tx
 import Cardano.Wallet.Primitive.Types.Tx.TxIn
     ( TxIn (..)
     )
+import Cardano.Wallet.Primitive.Types.ValidityIntervalExplicit
+    ( ValidityIntervalExplicit
+    )
+import Cardano.Wallet.Primitive.Types.WitnessCount
+    ( WitnessCount (WitnessCount)
+    , WitnessCountCtx
+    , toKeyRole
+    )
 import Cardano.Wallet.Read.Eras
-    ( babbage
+    ( conway
     , inject
-    )
-import Cardano.Wallet.Read.Primitive.Tx.Features.Certificates
-    ( anyEraCerts
-    )
-import Cardano.Wallet.Read.Primitive.Tx.Features.Inputs
-    ( fromShelleyTxIn
-    )
-import Cardano.Wallet.Read.Primitive.Tx.Features.Metadata
-    ( fromBabbageMetadata
-    )
-import Cardano.Wallet.Read.Primitive.Tx.Features.Mint
-    ( babbageMint
-    , fromLedgerScriptHash
-    )
-import Cardano.Wallet.Read.Primitive.Tx.Features.Outputs
-    ( fromBabbageTxOut
-    )
-import Cardano.Wallet.Read.Primitive.Tx.Features.Validity
-    ( afterShelleyValidityInterval
-    )
-import Cardano.Wallet.Read.Primitive.Tx.Features.Withdrawals
-    ( fromLedgerWithdrawals
     )
 import Cardano.Wallet.Read.Tx
     ( Tx (..)
@@ -92,34 +115,14 @@ import Cardano.Wallet.Read.Tx.Hash
 import Cardano.Wallet.Read.Tx.Withdrawals
     ( shelleyWithdrawals
     )
-
-import Cardano.Ledger.Babbage
-    ( AlonzoScript
-    , BabbageTxOut
-    )
-import Cardano.Wallet.Primitive.Types.AnyExplicitScripts
-    ( AnyExplicitScript (..)
-    )
-import Cardano.Wallet.Primitive.Types.TokenMapWithScripts
-    ( PlutusScriptInfo (PlutusScriptInfo)
-    , PlutusVersion (PlutusVersionV1, PlutusVersionV2, PlutusVersionV3)
-    , ReferenceInput (ReferenceInput)
-    , ScriptReference (ViaReferenceInput, ViaSpending)
-    , TokenMapWithScripts
-    )
-import Cardano.Wallet.Primitive.Types.ValidityIntervalExplicit
-    ( ValidityIntervalExplicit
-    )
-import Cardano.Wallet.Primitive.Types.WitnessCount
-    ( WitnessCount (WitnessCount)
-    , WitnessCountCtx
-    , toKeyRole
-    )
 import Control.Lens
     ( folded
     , (<&>)
     , (^.)
     , (^..)
+    )
+import Data.Foldable
+    ( toList
     )
 import Data.Map
     ( Map
@@ -131,12 +134,13 @@ import Data.Word
     ( Word32
     )
 import Ouroboros.Consensus.Cardano.Block
-    ( StandardBabbage
+    ( StandardConway
     )
 
 import qualified Cardano.Api.Shelley as Cardano
 import qualified Cardano.Ledger.Alonzo.Scripts as Alonzo
 import qualified Cardano.Ledger.Alonzo.Tx as Alonzo
+import qualified Cardano.Ledger.Api as Ledger
 import qualified Cardano.Ledger.BaseTypes as SL
 import qualified Cardano.Ledger.Core as Core
 import qualified Cardano.Ledger.Language as Language
@@ -149,8 +153,8 @@ import qualified Data.List as L
 import qualified Data.Map.Strict as Map
 import qualified Data.Set as Set
 
-fromBabbageTx
-    :: Alonzo.AlonzoTx (Cardano.ShelleyLedgerEra BabbageEra)
+fromConwayTx
+    :: Alonzo.AlonzoTx (Cardano.ShelleyLedgerEra ConwayEra)
     -> WitnessCountCtx
     -> ( W.Tx
        , [W.Certificate]
@@ -159,9 +163,9 @@ fromBabbageTx
        , Maybe ValidityIntervalExplicit
        , WitnessCount
        )
-fromBabbageTx tx witCtx =
+fromConwayTx tx witCtx =
     ( tx'
-    , anyEraCerts $ tx ^. bodyTxL . certsTxBodyL
+    , fmap fromConwayCerts . toList $ tx ^. bodyTxL . conwayCertsTxBodyL
     , assetsToMint
     , assetsToBurn
     , Just $ afterShelleyValidityInterval $ tx ^. bodyTxL.vldtTxBodyL
@@ -171,7 +175,7 @@ fromBabbageTx tx witCtx =
         (fromIntegral $ Set.size $ tx ^. witsTxL.bootAddrTxWitsL)
     )
   where
-    tx' = fromBabbageTx' tx
+    tx' = fromConwayTx' tx
     txId' = txId tx'
 
     anyScriptsFromTxOuts :: Map TokenPolicyId AnyExplicitScript
@@ -184,16 +188,16 @@ fromBabbageTx tx witCtx =
       where
         scriptWithHashIx
             :: Word32
-            -> BabbageTxOut StandardBabbage
+            -> BabbageTxOut StandardConway
             -> Maybe
                 ( ScriptReference
-                , ScriptHash StandardCrypto
-                , AlonzoScript StandardBabbage
+                , Ledger.ScriptHash StandardCrypto
+                , AlonzoScript StandardConway
                 )
         scriptWithHashIx ix txout =
-            snd (fromBabbageTxOut txout) <&> \script ->
+            snd (fromConwayTxOut txout) <&> \script ->
                 ( ViaReferenceInput (ReferenceInput (TxIn txId' ix))
-                , hashBabbageScript script
+                , hashConwayScript script
                 , script
                 )
 
@@ -205,15 +209,15 @@ fromBabbageTx tx witCtx =
             ]
 
     (assetsToMint, assetsToBurn) =
-        babbageMint
+        conwayMint
             (tx ^. bodyTxL.referenceInputsTxBodyL)
             (tx ^. bodyTxL.mintTxBodyL)
             (tx ^. witsTxL)
 
     fromLedgerToAnyScript
         :: ( ScriptReference
-           , ScriptHash StandardCrypto
-           , AlonzoScript StandardBabbage
+           , Ledger.ScriptHash StandardCrypto
+           , AlonzoScript StandardConway
            )
         -> (TokenPolicyId, AnyExplicitScript)
     fromLedgerToAnyScript (scriptRef, scriptH, script) =
@@ -228,38 +232,38 @@ fromBabbageTx tx witCtx =
                 PlutusExplicitScript
                     (PlutusScriptInfo
                         (toPlutusVer ver)
-                        (fromLedgerScriptHash $ hashBabbageScript script))
+                        (fromLedgerScriptHash $ hashConwayScript script))
                     scriptRef
 
         toPlutusVer Language.PlutusV1 = PlutusVersionV1
         toPlutusVer Language.PlutusV2 = PlutusVersionV2
         toPlutusVer Language.PlutusV3 = PlutusVersionV3
 
-    hashBabbageScript = Core.hashScript @(Cardano.ShelleyLedgerEra BabbageEra)
+    hashConwayScript = Core.hashScript @(Cardano.ShelleyLedgerEra ConwayEra)
 
-fromBabbageTx' :: Alonzo.AlonzoTx (Cardano.ShelleyLedgerEra BabbageEra) -> W.Tx
-fromBabbageTx' tx =
+fromConwayTx' :: Alonzo.AlonzoTx (Cardano.ShelleyLedgerEra ConwayEra) -> W.Tx
+fromConwayTx' tx =
     W.Tx
         { txId = W.Hash $ shelleyTxHash tx
         , txCBOR =
-            Just $ renderTxToCBOR $ inject babbage $ Tx tx
+            Just $ renderTxToCBOR $ inject conway $ Tx tx
         , fee =
-            Just $ Ledger.toWalletCoin $ tx ^. bodyTxL.feeTxBodyL
+            Just $ Ledger.toWalletCoin $ tx ^. bodyTxL . feeTxBodyL
         , resolvedInputs =
             (,Nothing) . fromShelleyTxIn
-                <$> tx ^.. bodyTxL.inputsTxBodyL.folded
+                <$> tx ^.. bodyTxL . inputsTxBodyL . folded
         , resolvedCollateralInputs =
             (,Nothing) . fromShelleyTxIn
-                <$> tx ^.. bodyTxL.collateralInputsTxBodyL.folded
+                <$> tx ^.. bodyTxL . collateralInputsTxBodyL . folded
         , outputs =
-            fst . fromBabbageTxOut <$> tx ^.. bodyTxL.outputsTxBodyL.folded
+            fst . fromConwayTxOut <$> tx ^.. bodyTxL . outputsTxBodyL . folded
         , collateralOutput =
             strictMaybeToMaybe $
-                fst . fromBabbageTxOut <$> tx ^. bodyTxL.collateralReturnTxBodyL
+                fst . fromConwayTxOut <$> tx ^. bodyTxL . collateralReturnTxBodyL
         , withdrawals =
             fromLedgerWithdrawals . shelleyWithdrawals $ tx
         , metadata =
-            fromBabbageMetadata <$> SL.strictMaybeToMaybe (tx ^. auxDataTxL)
+            fromConwayMetadata <$> SL.strictMaybeToMaybe (tx ^. auxDataTxL)
         , scriptValidity =
             Just $ case tx ^. isValidTxL of
                 Alonzo.IsValid True -> W.TxScriptValid
