+++ lib/primitive/lib/Cardano/Wallet/Primitive/Ledger/Read/Tx/Features/ExtraSigs.hs
@@ -8,7 +8,7 @@
 -- License: Apache-2.0
 --
 
-module Cardano.Wallet.Read.Primitive.Tx.Features.ExtraSigs
+module Cardano.Wallet.Primitive.Ledger.Read.Tx.Features.ExtraSigs
     ( extraSigs )
 
  where
