+++ lib/primitive/lib/Cardano/Wallet/Primitive/Ledger/Read/Tx/Features/Metadata.hs
@@ -8,7 +8,7 @@
 -- License: Apache-2.0
 --
 
-module Cardano.Wallet.Read.Primitive.Tx.Features.Metadata
+module Cardano.Wallet.Primitive.Ledger.Read.Tx.Features.Metadata
     ( getMetadata
     , fromShelleyMetadata
     , fromAllegraMetadata
