+++ lib/primitive/lib/Cardano/Wallet/Primitive/Ledger/Read/Tx/Features/Withdrawals.hs
@@ -7,7 +7,7 @@
 -- License: Apache-2.0
 --
 
-module Cardano.Wallet.Read.Primitive.Tx.Features.Withdrawals
+module Cardano.Wallet.Primitive.Ledger.Read.Tx.Features.Withdrawals
     ( getWithdrawals
     , fromLedgerWithdrawals
     )
@@ -35,8 +35,8 @@ import Data.Map.Strict
 import qualified Cardano.Ledger.Api as Ledger
 import qualified Cardano.Ledger.Coin as Ledger
 import qualified Cardano.Wallet.Primitive.Ledger.Convert as Ledger
+import qualified Cardano.Wallet.Primitive.Ledger.Read.Tx.Features.Certificates as Certificates
 import qualified Cardano.Wallet.Primitive.Types.Coin as W
-import qualified Cardano.Wallet.Read.Primitive.Tx.Features.Certificates as Certificates
 import qualified Data.Map as Map
 
 getWithdrawals :: EraFun Withdrawals (K (Maybe (Map RewardAccount Coin)))
