+++ lib/wallet/src/Cardano/Wallet/DB.hs
@@ -64,6 +64,9 @@ import Cardano.Wallet.DB.WalletState
     , WalletState (submissions)
     , updateSubmissions
     )
+import Cardano.Wallet.Primitive.Ledger.Read.Tx.Sealed
+    ( fromSealedTx
+    )
 import Cardano.Wallet.Primitive.Model
     ( Wallet
     , currentTip
@@ -110,9 +113,6 @@ import Cardano.Wallet.Primitive.Types.Tx.TxOut
     )
 import Cardano.Wallet.Read.Eras
     ( EraValue
-    )
-import Cardano.Wallet.Read.Primitive.Tx.Sealed
-    ( fromSealedTx
     )
 import Cardano.Wallet.Submissions.Submissions
     ( TxStatusMeta (..)
